feat: script to obtain licensing info for public repos
This script attempts to obtain licensing information for the public repos for a given Github
organisation, though it's heavily biased to Rust-based projects. A Ruby Gem called Licensee is used
for gathering the info on the repositories.

After analysis, it prints out a few tables with licensing info on:
* Crates: what's used in the Cargo manifest and mentioned in the README, both of which will appear
  on crates.io.
* Repos: what the Github API reports and what license files were detected by Licensee.
* Source files: the number of files that don't have a copyright notice and the years that appear in
  the notices.

The comments at the top of the script provide a pretty good guide for the setup required for it. The
reason I've not included a requirements.txt is because this is a generic repo for housing lots of
scripts.