feat: script to list ssh keys on digital ocean
For some reason this functionality doesn't appear to be in the `doctl` program and you have to make
a REST request for it. I always forget how to do it, so this little script lists the key name and
the ID number, which is necessary for use with things like Terraform.