feat: upload dvd images to s3 bucket
Just a basic script for uploading DVD image scans to my bucket on S3.

It produces a resized version of each file for the purposes of posting them on forums or other
places. Then it uploads each to an S3 bucket, using an ID that is generated for the DVD, which will
always be in the first line of the README file in the parent directory. Finally each of the links
get written out to a text file, so I can paste them easily.