feat: create new video by extracting segments
This is very useful for removing adverts from extended TV footage.

It will extract the segments defined in a `segments.txt` file in the directory where the script
executes.

Example `segments.txt`:
00:00:00 - 00:28:18
00:30:23 - 00:54:13
00:55:18 - 01:37:24
01:39:27 - 02:24:09
02:27:12 - 02:56:14

The segments are then joined together; the effect is to create a new video with the parts in between
each segment removed.

It can also operate on just one segment.